requirements.txt
librosa
noisereduce
soundfile
pandas
numpy
matplotlib
pyarabic
num2words
sentence-transformers
torch
tensorflow-macos # For macOS users
# tensorflow>=2.8.0  # Uncomment for non-macOS users
tqdm
jiwer
hf_xet
ipython
dask
